GH-6: Index debian Pre-Depends field for later use
- The initial ArtifactInfo created from the control file
is fed into Lucene for storage. At a later time, this info
is used to repopulate a new ArtifactInfo. The Pre-Depends
field was used before to create a Packages file, but it
was not stored in Lucene and therefore not available from
the ArtifactInfo when the Packages file was generated.

The Pre-Depends file is now stored in Lucene and available
in the generated Packages file